requirements.txt
amqp==2.6.0
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.7.0rc1
Django==2.1.2
django-braces==1.14.0
django-model-utils==4.0.0
django-private-chat==0.3.0
django-video-transcoding==0.1.0
djangorestframework==3.9.0
fffw==2.0.4
ffmpeg==1.4
ffmpy3==0.2.3
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
MediaInfo==0.0.8
numpy==1.19.0
oauthlib==3.1.0
pandas==1.0.5
Pillow==7.2.0
pycparser==2.20
pydub==0.24.1
PyJWT==1.7.1
pymediainfo==4.2.1
pypng==0.0.20
PyQRCode==1.2.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
qrcode==6.1
requests==2.24.0
requests-oauthlib==1.3.0
sendotp==0.0.4
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
spotipy==2.13.0
urllib3==1.25.9
vine==1.3.0
websockets==8.1
zipp==3.1.0
